+++ src/columnChart/columnChartVisual.ts
@@ -26,9 +26,8 @@
 import powerbi from "powerbi-visuals-api";
 import { legendInterfaces, axisInterfaces } from "powerbi-visuals-utils-chartutils";
 
-import VisualObjectInstance = powerbi.VisualObjectInstance;
-import EnumerateVisualObjectInstancesOptions = powerbi.EnumerateVisualObjectInstancesOptions;
 import IViewport = powerbi.IViewport;
+import ILocalizationManager = powerbi.extensibility.ILocalizationManager;
 
 import {
     MekkoChartVisualInitOptions,
@@ -46,16 +45,12 @@ export interface IColumnChart {
     getColumnsWidth(): number[];
 
     init(options: MekkoChartVisualInitOptions): void;
-    setData(dataViews: powerbi.DataView[], settingsModel: VisualFormattingSettingsModel, resized?: boolean): void;
+    setData(dataViews: powerbi.DataView[], settingsModel: VisualFormattingSettingsModel, isFormatMode: boolean, localizationManager: ILocalizationManager, resized?: boolean): void;
     calculateAxesProperties(options: MekkoCalculateScaleAndDomainOptions, settingsModel: VisualFormattingSettingsModel): IAxisProperties[];
     overrideXScale(xProperties: IAxisProperties): void;
     render(suppressAnimations: boolean, settingsModel: VisualFormattingSettingsModel): MekkoVisualRenderResult;
     calculateLegend(): LegendData;
     hasLegend(): boolean;
-    onClearSelection(): void;
-    enumerateObjectInstances?(
-        instances: VisualObjectInstance[],
-        options: EnumerateVisualObjectInstancesOptions): void;
     getVisualCategoryAxisIsScalar?(): boolean;
     getSupportedCategoryAxisType?(): string;
     getPreferredPlotArea?(
@@ -65,4 +60,5 @@ export interface IColumnChart {
     setFilteredData?(startIndex: number, endIndex: number): MekkoChartBaseData;
 
     getData?(): MekkoChartBaseData;
+    getAxisProperties?(): IAxisProperties[];
 }