+++ changelog.md
@@ -1,3 +1,20 @@
+## 3.7.0.0
+
+### New features
+* Added on-Object formatting API
+
+### Other
+* Updated dependencies
+
+## 3.6.3.0
+* Remove interactivity utils
+* Update packages
+* Increase selection visibility in high contrast mode
+* Use API 5.11.0
+
+## 3.6.2.0
+* Add high contrast mode support
+
 ## 3.6.1.0
 * Show context menu when right-clicking on empty spot inside the visual
 * Update powerbi utils and tools dependencies